Single upload per tick and switch to fp32.

+++ src/cpp/force_collide.cpp
@@ -8,15 +8,15 @@
 #include <cstdlib>
 
 struct CollideQuadNode {
-    double x0, y0, x1, y1;
+    float x0, y0, x1, y1;
     int dataIndex;
     int children[4];
-    double maxRadius;
+    float maxRadius;
 
     CollideQuadNode()
         : x0(0), y0(0), x1(0), y1(0), dataIndex(-1), children{-1, -1, -1, -1}, maxRadius(0) {}
 
-    CollideQuadNode(double x0_, double y0_, double x1_, double y1_)
+    CollideQuadNode(float x0_, float y0_, float x1_, float y1_)
         : x0(x0_), y0(y0_), x1(x1_), y1(y1_), dataIndex(-1), children{-1, -1, -1, -1}, maxRadius(0) {}
 };
 
@@ -26,34 +26,34 @@ private:
     std::function<double(const emscripten::val&, int, const emscripten::val&)> radiusFunc =
         [](const emscripten::val&, int, const emscripten::val&) { return 1.0; };
     bool radiusFuncIsCustom = false;
-    double radiusConstant = 1.0;
-    double strength = 1.0;
+    float radiusConstant = 1.0f;
+    float strength = 1.0f;
     int iterations = 1;
 
-    std::vector<double> radii;
-    std::vector<double> pxCache;
-    std::vector<double> pyCache;
+    std::vector<float> radii;
+    std::vector<float> pxCache;
+    std::vector<float> pyCache;
     std::vector<CollideQuadNode> quadtree;
     std::vector<std::pair<unsigned int, int>> mortonIndices;
     std::vector<int> sortedIndices;
 
-    double* nodeBuffer = nullptr;
-    double* radiusBufferPtr = nullptr;
+    float* nodeBuffer = nullptr;
+    float* radiusBufferPtr = nullptr;
     int nodeCount = 0;
     int radiusBufferCount = 0;
     static constexpr int stride = 4;
 
-    std::function<double()> random = []() {
-        return std::rand() / static_cast<double>(RAND_MAX);
+    std::function<float()> random = []() {
+        return static_cast<float>(std::rand()) / static_cast<float>(RAND_MAX);
     };
 
-    static unsigned int mortonCode(double x, double y, double x0, double y0, double span, unsigned int scaleFactor) {
-        double nx = span != 0 ? (x - x0) / span : 0.0;
+    static unsigned int mortonCode(float x, float y, float x0, float y0, float span, unsigned int scaleFactor) {
+        float nx = span != 0 ? (x - x0) / span : 0.0f;
         if (nx < 0.0) nx = 0.0;
-        if (nx > 0.999999) nx = 0.999999;
-        double ny = span != 0 ? (y - y0) / span : 0.0;
+        if (nx > 0.999999) nx = 0.999999f;
+        float ny = span != 0 ? (y - y0) / span : 0.0f;
         if (ny < 0.0) ny = 0.0;
-        if (ny > 0.999999) ny = 0.999999;
+        if (ny > 0.999999) ny = 0.999999f;
         unsigned int scaledX = static_cast<unsigned int>(nx * scaleFactor);
         unsigned int scaledY = static_cast<unsigned int>(ny * scaleFactor);
         unsigned int morton = 0;
@@ -71,21 +71,21 @@ private:
             radii.clear();
             return;
         }
-        radii.resize(nodeCount);
+            radii.resize(nodeCount);
         if (radiusFuncIsCustom) {
             for (int i = 0; i < nodeCount; ++i) {
                 double value = radiusFunc(nodes[i], i, nodes);
-                radii[i] = std::max(0.0, value);
+                radii[i] = static_cast<float>(std::max(0.0, value));
             }
         } else {
-            double constant = std::max(0.0, radiusConstant);
+            float constant = std::max(0.0f, radiusConstant);
             std::fill(radii.begin(), radii.end(), constant);
         }
     }
 
     int childIndexFor(const CollideQuadNode& quad, int dataIndex) const {
-        double midX = 0.5 * (quad.x0 + quad.x1);
-        double midY = 0.5 * (quad.y0 + quad.y1);
+        float midX = 0.5f * (quad.x0 + quad.x1);
+        float midY = 0.5f * (quad.y0 + quad.y1);
         int child = 0;
         if (pxCache[dataIndex] >= midX) child |= 1;
         if (pyCache[dataIndex] >= midY) child |= 2;
@@ -95,8 +95,8 @@ private:
     void subdivide(int quadIndex) {
         CollideQuadNode& quad = quadtree[quadIndex];
 
-        double midX = 0.5 * (quad.x0 + quad.x1);
-        double midY = 0.5 * (quad.y0 + quad.y1);
+        float midX = 0.5f * (quad.x0 + quad.x1);
+        float midY = 0.5f * (quad.y0 + quad.y1);
 
         int baseIndex = quadtree.size();
         quad.children[0] = baseIndex;
@@ -139,13 +139,13 @@ private:
         insertNode(quad.children[child], dataIndex);
     }
 
-    void updateMaxRadius(int quadIndex, const double* radiiValues) {
+    void updateMaxRadius(int quadIndex, const float* radiiValues) {
         CollideQuadNode& quad = quadtree[quadIndex];
         if (quad.children[0] == -1) {
             quad.maxRadius = (quad.dataIndex >= 0) ? radiiValues[quad.dataIndex] : 0.0;
             return;
         }
-        double maxR = 0.0;
+        float maxR = 0.0f;
         for (int i = 0; i < 4; ++i) {
             int child = quad.children[i];
             if (child >= 0) {
@@ -158,27 +158,27 @@ private:
         quad.maxRadius = maxR;
     }
 
-    void resolvePair(int nodeIndex, int otherIndex, const double* radiiValues) {
+    void resolvePair(int nodeIndex, int otherIndex, const float* radiiValues) {
         if (nodeIndex == otherIndex || otherIndex < nodeIndex) return;
 
-        double dx = pxCache[nodeIndex] - pxCache[otherIndex];
-        double dy = pyCache[nodeIndex] - pyCache[otherIndex];
-        double r = radiiValues[nodeIndex] + radiiValues[otherIndex];
+        float dx = pxCache[nodeIndex] - pxCache[otherIndex];
+        float dy = pyCache[nodeIndex] - pyCache[otherIndex];
+        float r = radiiValues[nodeIndex] + radiiValues[otherIndex];
         if (r <= 0.0) return;
 
-        double r2 = r * r;
-        double l2 = dx * dx + dy * dy;
+        float r2 = r * r;
+        float l2 = dx * dx + dy * dy;
         if (l2 >= r2) return;
 
-        if (l2 == 0.0) {
-            dx = (random() - 0.5) * 1e-6;
-            dy = (random() - 0.5) * 1e-6;
+        if (l2 == 0.0f) {
+            dx = (random() - 0.5f) * 1e-6f;
+            dy = (random() - 0.5f) * 1e-6f;
             l2 = dx * dx + dy * dy;
-            if (l2 == 0.0) return;
+            if (l2 == 0.0f) return;
         }
 
-        double l = std::sqrt(l2);
-        double adjustment = (l - r) / l * strength;
+        float l = std::sqrt(l2);
+        float adjustment = (l - r) / l * strength;
         dx *= adjustment;
         dy *= adjustment;
 
@@ -191,13 +191,13 @@ private:
         nodeBuffer[offsetOther + 3] += dy;
     }
 
-    void visit(int quadIndex, int nodeIndex, const double* radiiValues) {
+    void visit(int quadIndex, int nodeIndex, const float* radiiValues) {
         CollideQuadNode& quad = quadtree[quadIndex];
-        double r = radiiValues[nodeIndex] + quad.maxRadius;
-        if (r <= 0.0) return;
+        float r = radiiValues[nodeIndex] + quad.maxRadius;
+        if (r <= 0.0f) return;
 
-        double px = pxCache[nodeIndex];
-        double py = pyCache[nodeIndex];
+        float px = pxCache[nodeIndex];
+        float py = pyCache[nodeIndex];
         if (quad.x0 > px + r || quad.x1 < px - r ||
             quad.y0 > py + r || quad.y1 < py - r) {
             return;
@@ -224,7 +224,7 @@ public:
     void force(double /*alpha*/) {
         if (nodeBuffer == nullptr || nodeCount == 0) return;
 
-        const double* radiiValues = nullptr;
+        const float* radiiValues = nullptr;
         if (radiusBufferPtr && radiusBufferCount >= nodeCount) {
             radiiValues = radiusBufferPtr;
         } else {
@@ -232,7 +232,7 @@ public:
                 computeRadii();
             }
             if (static_cast<int>(radii.size()) < nodeCount) {
-                radii.resize(nodeCount, std::max(0.0, radiusConstant));
+                radii.resize(nodeCount, std::max(0.0f, radiusConstant));
             }
             radiiValues = radii.data();
         }
@@ -243,19 +243,19 @@ public:
         sortedIndices.resize(nodeCount);
 
         for (int iter = 0; iter < iterations; ++iter) {
-            double x0 = std::numeric_limits<double>::infinity();
-            double y0 = std::numeric_limits<double>::infinity();
-            double x1 = -std::numeric_limits<double>::infinity();
-            double y1 = -std::numeric_limits<double>::infinity();
+            float x0 = std::numeric_limits<float>::infinity();
+            float y0 = std::numeric_limits<float>::infinity();
+            float x1 = -std::numeric_limits<float>::infinity();
+            float y1 = -std::numeric_limits<float>::infinity();
 
             for (int i = 0; i < nodeCount; ++i) {
                 int offset = i * stride;
-                double x = nodeBuffer[offset];
-                double y = nodeBuffer[offset + 1];
-                double vx = nodeBuffer[offset + 2];
-                double vy = nodeBuffer[offset + 3];
-                double px = x + vx;
-                double py = y + vy;
+                float x = nodeBuffer[offset];
+                float y = nodeBuffer[offset + 1];
+                float vx = nodeBuffer[offset + 2];
+                float vy = nodeBuffer[offset + 3];
+                float px = x + vx;
+                float py = y + vy;
                 pxCache[i] = px;
                 pyCache[i] = py;
                 if (px < x0) x0 = px;
@@ -265,20 +265,20 @@ public:
             }
 
             if (!std::isfinite(x0)) {
-                x0 = y0 = x1 = y1 = 0.0;
+                x0 = y0 = x1 = y1 = 0.0f;
             }
 
-            double dx = x1 - x0;
-            double dy = y1 - y0;
-            if (dx == 0.0) dx = 1.0;
-            if (dy == 0.0) dy = 1.0;
+            float dx = x1 - x0;
+            float dy = y1 - y0;
+            if (dx == 0.0f) dx = 1.0f;
+            if (dy == 0.0f) dy = 1.0f;
             x0 -= dx * 0.1;
             x1 += dx * 0.1;
             y0 -= dy * 0.1;
             y1 += dy * 0.1;
 
-            double span = std::max(x1 - x0, y1 - y0);
-            if (span == 0.0) span = 1.0;
+            float span = std::max(x1 - x0, y1 - y0);
+            if (span == 0.0f) span = 1.0f;
             unsigned int scaleFactor = 65535u;
 
             for (int i = 0; i < nodeCount; ++i) {
@@ -320,12 +320,12 @@ public:
     }
 
     void setNodeBuffer(uintptr_t ptr, int count) {
-        nodeBuffer = (ptr && count > 0) ? reinterpret_cast<double*>(ptr) : nullptr;
+        nodeBuffer = (ptr && count > 0) ? reinterpret_cast<float*>(ptr) : nullptr;
         nodeCount = count;
     }
 
     void setRadiusBuffer(uintptr_t ptr, int count) {
-        radiusBufferPtr = (ptr && count > 0) ? reinterpret_cast<double*>(ptr) : nullptr;
+        radiusBufferPtr = (ptr && count > 0) ? reinterpret_cast<float*>(ptr) : nullptr;
         radiusBufferCount = count;
     }
